Refactor AdminPanels
