+++ python/lsst/ts/rubintv/templates/admin.jinja
@@ -25,8 +25,7 @@
 {% block footer_scripts %}
 <script src="{{ url_for('static', path='assets/admin.js') }}"></script>
 <script>
-    window.APP_DATA.redisGetURL = "{{ url_for('redis_get')}}"
     window.APP_DATA.admin = {{ admin|tojson }};
-    window.APP_DATA.baseURL = "{{ request.base_url }}";
+    window.APP_DATA.baseUrl = "{{ request.base_url }}";
 </script>
 {% endblock footer_scripts %}
