+++ src/sass/admin.sass
@@ -1,5 +1,6 @@
 #admin-panels
     width: min-content
+    margin-bottom: 2em
 .admin-panel
     display: flex
     flex-direction: column
